refactor(update-imports): update the imports to point to the refactored routerUtils
