+++ src/store/index.js
@@ -1,37 +1,29 @@
+/* eslint-disable import/no-import-module-exports */
 /* eslint-disable global-require */
 /* eslint-disable no-underscore-dangle */
 import { applyMiddleware, createStore, compose } from 'redux'
 import createSagaMiddleware from 'redux-saga'
 import thunk from 'redux-thunk'
 import { router5Middleware } from 'redux-router5'
-
 import SagaManager from './sagaManager'
 import reducer from './reducer'
 
 const Store = (router, initialState = {}) => {
   const sagaMiddleware = createSagaMiddleware()
 
-  const middleware = [
-    router5Middleware(router),
-    thunk,
-    sagaMiddleware,
-  ]
+  const middleware = [router5Middleware(router), thunk, sagaMiddleware]
 
-  const storeEnhancers = [
-    applyMiddleware(...middleware),
-  ]
+  const storeEnhancers = [applyMiddleware(...middleware)]
 
   if (window.__REDUX_DEVTOOLS_EXTENSION__) {
-    storeEnhancers.push(window.__REDUX_DEVTOOLS_EXTENSION__({
-      shouldHotReload: false,
-    }))
+    storeEnhancers.push(
+      window.__REDUX_DEVTOOLS_EXTENSION__({
+        shouldHotReload: false,
+      }),
+    )
   }
 
-  const store = createStore(
-    reducer,
-    initialState,
-    compose(...storeEnhancers),
-  )
+  const store = createStore(reducer, initialState, compose(...storeEnhancers))
 
   router.setDependency('store', store)
   SagaManager.startSagas(sagaMiddleware, router)
